Extract the TransactionDetails comp to shared dir
